company-smart-access-system/services/gateway-service/requirements.txt
fastapi
uvicorn[standard]
httpx
python-jose[cryptography]
